Refactor string helper functions
