+++ src/frontend/lib/functions/String.tsx
